+++ examples/direct_mutation.rs
@@ -37,7 +37,7 @@ fn main() {
 
 // The player has different abilities depending on the color mode.
 // Yellow mode is its own thing, for example; not just red and green at the same time.
-#[derive(State, Clone, PartialEq, Eq, Default)]
+#[derive(State, Component, Clone, PartialEq, Eq, Default)]
 #[state(log_flush)]
 struct ColorMode {
     red: bool,